feat : ParticleSystem 에 업데이트용 ComputeShader 적용 중

+++ OutputFile/content/shader/SetColor.fx
@@ -15,7 +15,7 @@ RWTexture2D<float4> g_TargetTex : register(u0);
 [numthreads(32, 32, 1)]
 void CS_SetColor(uint3 id : SV_DispatchThreadID)
 {
-    if (TEX_WIDTH <= id.x || TEX_HEIGHT <= id.y)
+    if (TEX_WIDTH <= (int) id.x || TEX_HEIGHT <= (int) id.y)
     {
         return;
     }
